+++ src/noesis_lists.erl
@@ -18,6 +18,8 @@
 % API
 -export([
   group_by/2,
+  pfilter/2,
+  pfilter/3,
   pmap/2,
   pmap/3,
   split/2
@@ -35,9 +37,28 @@ group_by(Fun, List) ->
   end, dict:new(), [{Fun(X), X} || X <- List]),
   dict:to_list(Dict).
 
-% @doc Delegates to {@link pmap/3} and sets default options.<br /><br />
+% @doc Delegates to {@link pfilter/3} and uses default options.<br /><br />
 %      <strong>Default Options</strong><br />
 %      <pre><code>[{retain_order, true}, {parallelism, round(erlang:system_info(schedulers) * 1.5)}]</code></pre>
+-spec pfilter(fun((T) -> boolean()), [T]) -> [T].
+pfilter(Fun, List) ->
+  pfilter(Fun, List, []).
+
+-spec pfilter(fun((T) -> boolean()), [T], noesis_proplists:proplist(atom(), term())) -> [T].
+pfilter(_Fun, [], _Options) ->
+  [];
+pfilter(Fun, List, Options) ->
+  case noesis_proplists:get_value(parallelism, Options, round(erlang:system_info(schedulers) * 1.5)) of
+    1 -> lists:filter(Fun, List);
+    Parallelism ->
+      List2 = parallel_init(Fun, Parallelism, List),
+      pfilter_extract_results(List2, Options)
+  end.
+
+% @doc Delegates to {@link pmap/3} and uses default options.<br /><br />
+%      <strong>Default Options</strong><br />
+%      <pre><code>[{retain_order, true}, {parallelism, round(erlang:system_info(schedulers) * 1.5)}]</code></pre>
+-spec pmap(fun((A) -> B), [A]) -> [B].
 pmap(Fun, List) ->
   pmap(Fun, List, []).
 
@@ -52,10 +73,7 @@ pmap(Fun, List, Options) ->
   case noesis_proplists:get_value(parallelism, Options, round(erlang:system_info(schedulers) * 1.5)) of
     1 -> lists:map(Fun, List);
     Parallelism ->
-      Ref = make_ref(),
-      {Chunks, Rest} = split(Parallelism, List),
-      Index = parallel_run(Fun, Ref, 1, Chunks),
-      List2 = parallel_run_and_gather(Fun, Ref, Rest, length(List), Index, [], 0),
+      List2 = parallel_init(Fun, Parallelism, List),
       pmap_extract_results(List2, Options)
   end.
 
@@ -75,6 +93,13 @@ split(_N, List) ->
 
 % Private
 
+-spec parallel_init(fun(), pos_integer(), list()) -> [{pos_integer(), term(), term()}].
+parallel_init(Fun, Parallelism, List) ->
+  Ref = make_ref(),
+  {Chunks, Rest} = split(Parallelism, List),
+  Index = parallel_run(Fun, Ref, 1, Chunks),
+  parallel_run_and_gather(Fun, Ref, Rest, length(List), Index, [], 0).
+
 -spec parallel_run(fun(), reference(), pos_integer(), list()) -> pos_integer().
 parallel_run(Fun, Ref, StartIndex, List) ->
   Parent = self(),
@@ -86,7 +111,7 @@ parallel_run(Fun, Ref, StartIndex, List) ->
     Index + 1
   end, StartIndex, List).
 
--spec parallel_run_and_gather(fun(), reference(), list(), non_neg_integer(), pos_integer(), list(), non_neg_integer()) -> list().
+-spec parallel_run_and_gather(fun(), reference(), list(), non_neg_integer(), pos_integer(), list(), non_neg_integer()) -> [{pos_integer(), term(), term()}].
 parallel_run_and_gather(_Fun, _Ref, [], ResultLength, _Index, Acc, AccLength) when ResultLength =:= AccLength ->
   Acc;
 parallel_run_and_gather(Fun, Ref, [], ResultLength, Index, Acc, AccLength) ->
@@ -105,7 +130,7 @@ parallel_run_and_gather(Fun, Ref, [Chunk|Rest], ResultLength, Index, Acc, AccLen
       parallel_run_and_gather(Fun, Ref, Rest, ResultLength, Index2, Acc2, AccLength2)
   end.
 
--spec pmap_extract_results([{pos_integer(), A}], noesis_proplists:proplist(atom(), term())) -> [A].
+-spec pmap_extract_results([{pos_integer(), term(), A}], noesis_proplists:proplist(atom(), term())) -> [A].
 pmap_extract_results(List, Options) ->
   List2 = case noesis_proplists:get_value(retain_order, Options, true) of
     true -> lists:keysort(1, List);
@@ -113,3 +138,11 @@ pmap_extract_results(List, Options) ->
   end,
   {_Keys, _Items, Values} = lists:unzip3(List2),
   Values.
+
+-spec pfilter_extract_results([{pos_integer(), T, boolean()}], noesis_proplists:proplist(atom(), term())) -> [T].
+pfilter_extract_results(List, Options) ->
+  List2 = case noesis_proplists:get_value(retain_order, Options, true) of
+    true -> lists:keysort(1, List);
+    false -> List
+  end,
+  [Item || {_Key, Item, Value} <- List2, Value =:= true].