+++ test/noesis_lists_test.erl
@@ -18,17 +18,28 @@ group_by_test() ->
   ?assertEqual([{x, [{x, 1}, {x, 3}]}, {y, [{y, 2}]}], noesis_lists:group_by(fun({K, _V}) -> K end, [{x, 1}, {y, 2}, {x, 3}])),
   ?assertEqual([], noesis_lists:group_by(fun(_V) -> x end, [])).
 
+pfilter_test() ->
+  FunA = fun(X) -> X rem 2 =:= 0 end,
+  ListA = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 66, 99, 6, 3, 9000],
+  ?assertEqual(lists:filter(FunA, ListA), noesis_lists:pfilter(FunA, ListA)),
+  ?assertEqual(lists:sum(lists:filter(FunA, ListA)), lists:sum(noesis_lists:pfilter(FunA, ListA))),
+  ?assertEqual(lists:sum(noesis_lists:pfilter(FunA, ListA)), lists:sum(noesis_lists:pfilter(FunA, ListA, [{retain_order, false}]))),
+  ?assertEqual([], noesis_lists:pfilter(FunA, [])),
+  ?assertEqual(lists:filter(FunA, ListA), noesis_lists:pfilter(FunA, ListA, [{parallelism, 1}])),
+  ?assertEqual(lists:filter(FunA, ListA), noesis_lists:pfilter(FunA, ListA, [{parallelism, 2}])).
+
 pmap_test() ->
   FunA = fun(X) -> X + 1 end,
   ListA = [1, 2, 10, 66, 99, 6, 3, 9000],
   ?assertEqual(lists:map(FunA, ListA), noesis_lists:pmap(FunA, ListA)),
+  ?assertEqual(lists:sum(lists:map(FunA, ListA)), lists:sum(noesis_lists:pmap(FunA, ListA))),
   ?assertEqual(lists:sum(noesis_lists:pmap(FunA, ListA)), lists:sum(noesis_lists:pmap(FunA, ListA, [{retain_order, false}]))),
+  ?assertEqual(lists:map(FunA, ListA), noesis_lists:pmap(FunA, ListA, [{parallelism, 1}])),
+  ?assertEqual(lists:map(FunA, ListA), noesis_lists:pmap(FunA, ListA, [{parallelism, 2}])),
   FunB = fun({_K, V}) -> V end,
   ListB = [{x, <<"hello">>}, {y, <<"test">>}, {z, <<"foo">>}],
   ?assertEqual(lists:map(FunB, ListB), noesis_lists:pmap(FunB, ListB)),
-  ?assertEqual([], noesis_lists:pmap(FunB, [])),
-  ?assertEqual(lists:map(FunA, ListA), noesis_lists:pmap(FunA, ListA, [{parallelism, 1}])),
-  ?assertEqual(lists:map(FunA, ListA), noesis_lists:pmap(FunA, ListA, [{parallelism, 2}])).
+  ?assertEqual([], noesis_lists:pmap(FunB, [])).
 
 split_test() ->
   List = [1, 2, 3, 4, 5, 6, 7, 8, 9],