+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.2
 
 * Switch from [Rebar](https://github.com/rebar/rebar) to [erlang.mk](https://github.com/ninenines/erlang.mk)
-* `noesis_lists`: Add `pmap/{2,3}`, `split/2`
+* `noesis_lists`: Add ``pfilter/{2,3}``, `pmap/{2,3}`, `split/2`
 
 ## 0.1